Refactor python SDK, implement support for M2M Auth
